+++ service/src/main/kotlin/io/provenance/explorer/grpc/v1/ValidatorGrpcClient.kt
@@ -163,4 +163,9 @@ class ValidatorGrpcClient(channelUri: URI) {
         slashingClient.signingInfos(SlashingOuterClass.QuerySigningInfosRequest.getDefaultInstance()).infoList
 
     fun getSlashingParams() = slashingClient.params(SlashingOuterClass.QueryParamsRequest.newBuilder().build())
+
+    fun getDistParams() = distClient.params(DistOuterClass.QueryParamsRequest.newBuilder().build())
+
+    fun getStakingParams() = stakingClient.params(StakingOuterClass.QueryParamsRequest.newBuilder().build())
+
 }
